S10, fixed issues that had arrisen in tests due to files being moved around, updated readme to refllect current itteration of app while providing more detail on the packages and libraries used, moved non essencial component from production to dev inorder to prevent false alarm protentially blocking the building of the application, impliment comments through out the application inorder to more clearly communicate how each part functions
