+++ package.json
@@ -23,9 +23,7 @@
     "react-bootstrap": "^2.4.0",
     "react-dom": "^18.2.0",
     "react-router": "^6.3.0",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.1",
-    "react-xarrows": "^2.0.2"
+    "react-router-dom": "^6.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -50,6 +48,7 @@
     ]
   },
   "devDependencies": {
+    "react-scripts": "5.0.1",
     "@testing-library/dom": "^8.16.1",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
