+++ src/__tests__/App.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '../setupTests';
-import App from '../Components/App';
+import App from '../Components/App/App';
 
 describe("<App />", () => {
   test("Renders <App /> component correctly", () => {
