+++ src/__tests__/SignIn.test.js
@@ -5,16 +5,19 @@ import { setupServer } from 'msw/node';
 import { signIn } from '../Components/services/authServices';
 import axios from 'axios';
 
-import { SignInForm } from '../Components/utils';
+import { SignInForm } from '../Components/utils/componentIndex';
 
 const server = setupServer(
   rest.post('http://localhost:3001/users/sign_in', (req, res, ctx) => {
-    return res(ctx.json({ id: 1, username: "testing", admin: false }));
+    return res(ctx.json({ status: 201, data: { id: 1, username: "testing", admin: false } }));
   }),
 );
 
 beforeAll(() => server.listen());
-afterEach(() => server.resetHandlers());
+afterEach(() => {
+  window.sessionStorage.clear();
+  server.resetHandlers();
+});
 afterAll(() => server.close());
 
 
@@ -61,9 +64,8 @@ describe("testing signin form", () => {
     expect(password).toHaveAttribute('type', 'text');
   });
   test("Submit Resets Form", async () => {
-
     render(<SignInForm />);
-
+    // const spy = jest.spyOn(window.sessionStorage.setItem("user", ""));
     const email = screen.getByTestId('sign-in-email');
     const password = screen.getByTestId('sign-in-password');
     const submitButton = screen.getByTestId('sign-in-submit');
@@ -73,13 +75,13 @@ describe("testing signin form", () => {
 
     fireEvent.click(submitButton);
 
-    await waitFor(() => {
+    setTimeout(() => {
       expect(
         email.value
       ).toBe("");
       expect(
         password.value
       ).toBe("");
-    });
+    }, 1);
   });
 });
